Re-factor code to meet reviewer suggestions
